Added ruby app definition to docker-compose.yaml
We defined the ruby app into the yaml config file of
docker-compose. Now with a simple
```docker-compose up -d```
we will have a ruby app that connects to a redis automatized.
This is a first approach of our final goal. Now we have the right
container images and we are ready to try to run this images into
kubernetes.