Run a simple redis into kubernetes
With this file we will be able to run a redis-server into k8s,
allways using kubectl to interact with it (kubernetes).
```kubectl create -f ./run-redis.yaml```
It creates a deployment named redis-hello-world, that creates a pod
(container) that runs redis from the official redis image published
in docker hub.