Expose hello-world deployment as a k8s service.
Now if you expose this deployment as a k8s service and run:
```kubectl get services```
You are able to see the ipaddress where to do the requests to our
hello-world app. The app has 2 replicas, so we have 2 containers,
and it is load balanced by k8s. We achieved our goal.